requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.5.0
idna==3.10
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
multidict==6.1.0
packaging==24.1
pillow==11.0.0
pluggy==1.5.0
propcache==0.2.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
sqlparse==0.5.1
twilio==9.3.6
urllib3==2.2.3
yarl==1.16.0
